:sparkles: started implementing Lsp source

+++ compleet-client/src/bindings/api/buffer.rs
@@ -8,7 +8,7 @@ use mlua::{
 /// Binding to `vim.api.nvim_buf_attach`.
 pub fn buf_attach(
     lua: &Lua,
-    bufnr: u32,
+    bufnr: u16,
     send_buffer: bool,
     opts: Table,
 ) -> LuaResult<bool> {
@@ -21,7 +21,7 @@ pub fn buf_attach(
 
 #[allow(dead_code)]
 /// Binding to `vim.api.nvim_buf_call`.
-pub fn buf_call(lua: &Lua, bufnr: u32, fun: LuaFunction) -> LuaResult<()> {
+pub fn buf_call(lua: &Lua, bufnr: u16, fun: LuaFunction) -> LuaResult<()> {
     super::api(lua)?
         .get::<&str, LuaFunction>("nvim_buf_call")?
         .call((bufnr, fun))
@@ -30,7 +30,7 @@ pub fn buf_call(lua: &Lua, bufnr: u32, fun: LuaFunction) -> LuaResult<()> {
 /// Binding to `vim.api.nvim_buf_get_lines`.
 pub fn buf_get_lines(
     lua: &Lua,
-    bufnr: u32,
+    bufnr: u16,
     start: u32,
     end: i32,
     strict_indexing: bool,
@@ -46,7 +46,7 @@ pub fn buf_get_lines(
 /// Binding to `vim.api.nvim_buf_get_lines`.
 pub fn buf_get_option<'lua, V: FromLua<'lua>>(
     lua: &'lua Lua,
-    bufnr: u32,
+    bufnr: u16,
     name: &str,
 ) -> LuaResult<V> {
     super::api(lua)?
@@ -57,7 +57,7 @@ pub fn buf_get_option<'lua, V: FromLua<'lua>>(
 /// Binding to `vim.api.nvim_buf_set_lines`.
 pub fn buf_set_lines(
     lua: &Lua,
-    bufnr: u32,
+    bufnr: u16,
     start: u32,
     end: i32,
     strict_indexing: bool,
@@ -75,7 +75,7 @@ pub fn buf_set_lines(
 /// Binding to `vim.api.nvim_buf_set_text`.
 pub fn buf_set_text(
     lua: &Lua,
-    bufnr: u32,
+    bufnr: u16,
     start_row: u32,
     start_col: u32,
     end_row: u32,
