+++ compleet-client/src/bindings/api/global.rs
@@ -18,7 +18,7 @@ pub fn create_user_command(
 }
 
 /// Binding to `vim.api.nvim_create_buf`.
-pub fn create_buf(lua: &Lua, listed: bool, scratch: bool) -> LuaResult<u32> {
+pub fn create_buf(lua: &Lua, listed: bool, scratch: bool) -> LuaResult<u16> {
     super::api(lua)?
         .get::<&str, LuaFunction>("nvim_create_buf")?
         .call((listed, scratch))
@@ -46,7 +46,7 @@ pub fn echo(
 }
 
 /// Binding to `vim.api.nvim_get_current_buf`
-pub fn get_current_buf(lua: &Lua) -> LuaResult<u32> {
+pub fn get_current_buf(lua: &Lua) -> LuaResult<u16> {
     super::api(lua)?.get::<&str, LuaFunction>("nvim_get_current_buf")?.call(())
 }
